Fix some of the unittests for python 3. AltitudeRadio tests still fail.

+++ analysis_engine/library.py
@@ -4692,7 +4692,7 @@ def blend_parameters(params, offset=0.0, frequency=1.0, small_slice_duration=4, 
 
     tol_mask = None
     if tolerance:
-        test_array = np.ma.zeros((len(params), len(params[0].array) * min_ip_freq / params[0].frequency))
+        test_array = np.ma.zeros((len(params), int(len(params[0].array) * min_ip_freq / params[0].frequency)))
         for n, p in enumerate(params):
             test_array[n, :] = resample(p.array, p.frequency, min_ip_freq)
         tol_mask = np.ma.masked_greater(np.ma.ptp(test_array, axis=0), tolerance)
@@ -4726,7 +4726,7 @@ def blend_parameters(params, offset=0.0, frequency=1.0, small_slice_duration=4, 
         # collation.
         p_valid_slices.append(slices_multiply(nts, min_ip_freq / param.frequency))
 
-    all_masks = np.array([np.ma.getmaskarray(p.array)[::p.frequency // min_ip_freq] for p in params])
+    all_masks = np.array([np.ma.getmaskarray(p.array)[::int(p.frequency // min_ip_freq)] for p in params])
     num_valid = len(params) - np.sum(all_masks, axis=0)
 
     if validity == 'all':
@@ -5417,14 +5417,14 @@ def pin_to_ground(array, good_slices, fast_slices, hz=1.0):
         for n, sl in enumerate(good_slices):
             if is_index_within_slice(f.start, sl):
                 # go_fast starts in the slice
-                begin_step = nearest_step(array[f.start])
+                begin_step = nearest_step(array[int(f.start)])
                 array[sl] -=  begin_step
                 begin_peak_idx = sl.stop - 1
                 begin_peak = array[begin_peak_idx]
                 
             elif is_index_within_slice(f.stop, sl):
                 # go_fast stops in the slice
-                end_step = nearest_step(array[f.stop])
+                end_step = nearest_step(array[int(f.stop)])
                 array[sl] -=  end_step
                 end_peak_idx = sl.start
                 end_peak = array[end_peak_idx]
